pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fc-engineering-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyperclip>=1.11.0",
    "pyyaml>=6.0.3",
    "textual>=8.2.3",
]

[project.scripts]
fc-engineer = "main:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
